@@ -13,38 +13,39 @@
         "lint": "eslint ."
     },
     "dependencies": {
-        "@openapitools/openapi-generator-cli": "^2.5.1",
+        "@openapitools/openapi-generator-cli": "^2.6.0",
         "@types/headroom.js": "^0.12.2",
-        "@types/node": "^18.7.14",
-        "@types/react": "^18.0.18",
-        "@types/react-dom": "^18.0.6",
+        "@types/node": "^20.4.2",
+        "@types/react": "^18.2.15",
+        "@types/react-dom": "^18.2.7",
         "@types/react-router-dom": "^5.3.3",
         "aos": "^3.0.0-beta.6",
-        "argon-design-system-react": "^1.1.0",
         "babel-eslint": "^10.1.0",
-        "babel-jest": "^29.0.1",
+        "babel-jest": "^29.6.1",
+        "bootstrap": "^5.2.6",
         "cross-env": "^7.0.3",
         "flag-icon-css": "^4.1.7",
         "headroom.js": "^0.12.0",
-        "i18next": "^21.9.1",
+        "i18next": "^23.2.11",
+        "node-sass": "^9.0.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-i18next": "^11.18.5",
-        "react-markdown": "^8.0.3",
-        "react-router-dom": "^6.3.0",
+        "react-i18next": "^13.0.2",
+        "react-markdown": "^8.0.7",
+        "react-router-dom": "^6.14.1",
         "react-scripts": "^5.0.1",
-        "reactstrap": "^9.1.4",
-        "redux": "^4.2.0",
-        "typescript": "^4.8.2",
+        "reactstrap": "^9.2.0",
+        "redux": "^4.2.1",
+        "typescript": "^5.1.6",
         "universal-cookie": "^4.0.4"
     },
     "devDependencies": {
         "@types/aos": "^3.0.4",
-        "@typescript-eslint/eslint-plugin": "^4.29.2",
-        "@typescript-eslint/parser": "^4.29.2",
-        "eslint": "^7.32.0",
-        "eslint-plugin-import": "^2.24.0",
-        "rimraf": "^3.0.2"
+        "@typescript-eslint/eslint-plugin": "^6.2.0",
+        "@typescript-eslint/parser": "^6.2.0",
+        "eslint": "^8.45.0",
+        "eslint-plugin-import": "^2.27.5",
+        "rimraf": "^5.0.1"
     },
     "browserslist": {
         "production": [
