@@ -1,10 +1,10 @@
-import React    from 'react';
-import ReactDOM from 'react-dom';
-import App      from './components/App';
+import React from 'react';
+import ReactDOM from "react-dom/client";
+import App from './components/App';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+root.render(
     <React.StrictMode>
         <App/>
-    </React.StrictMode>,
-    document.getElementById('root')
+    </React.StrictMode>
 );
